OpenGL POG

+++ include/Simulator.hpp
@@ -2,17 +2,17 @@
 #include "Scene.hpp"
 
 struct Simulator {
-    virtual void update(float elapsed, Scene& scene) = 0;
+    virtual void update(double elapsed, Scene& scene) = 0;
 };
 
 struct SequentialSimulator : Simulator {
     SequentialSimulator();
 
-    void update(float elapsed, Scene& scene) override;
+    void update(double elapsed, Scene& scene) override;
 };
 
 struct ParallelSimulator : Simulator {
     ParallelSimulator();
 
-    void update(float elapsed, Scene& scene) override;
+    void update(double elapsed, Scene& scene) override;
 };
